Continue code clean-up: harden runtime inputs
- Size AppMessage buffers from the expected payload and retry with SDK minimums on open failure.

- Guard shared date, time, and display helpers against rare NULL inputs.

- Make settings persistence best-effort with explicit read/write diagnostics and README documentation.

- Validate raw weather values in the weather module and keep weather glyph scaling on helper_scale_round().

+++ src/modules/weather.h
@@ -20,10 +20,10 @@ void weather_module_refresh(
     uint8_t temp_unit,
     const VisualPalette* palette);
 void weather_module_set_temperature(
-    int16_t celsius_tenths,
+    int celsius_tenths,
     uint8_t temp_unit,
     const VisualPalette* palette);
 void weather_module_set_condition(
-    int16_t weather_condition,
+    int weather_condition,
     uint8_t temp_unit,
     const VisualPalette* palette);
